+++ backend/src/core/services/index.ts
@@ -9,3 +9,4 @@ export * from './api-key.service'
 export * from './auth.service'
 export * from './protected.service'
 export * from './parse-csv.service'
+export * from './multipart-upload.service'
